refactor : restruct test packages
