+++ src/test/java/kitchenpos/AcceptanceTest.java
@@ -1,0 +1,24 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+
+import io.restassured.RestAssured;
+import kitchenpos.utils.DatabaseCleanup;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+	@LocalServerPort
+	int port;
+
+	@Autowired
+	private DatabaseCleanup databaseCleanup;
+
+	@BeforeEach
+	public void setUp() {
+		RestAssured.port = port;
+		databaseCleanup.execute();
+	}
+}
