--- src/test/java/kitchenpos/acceptance/AcceptanceTest.java
@@ -1,24 +1,0 @@
-package kitchenpos.acceptance;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-
-import io.restassured.RestAssured;
-import kitchenpos.acceptance.utils.DatabaseCleanup;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class AcceptanceTest {
-	@LocalServerPort
-	int port;
-
-	@Autowired
-	private DatabaseCleanup databaseCleanup;
-
-	@BeforeEach
-	public void setUp() {
-		RestAssured.port = port;
-		databaseCleanup.execute();
-	}
-}
