--- src/test/java/kitchenpos/ui/WebMvcTestConfiguration.java
@@ -1,31 +1,0 @@
-package kitchenpos.ui;
-
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class WebMvcTestConfiguration {
-	@Autowired
-	protected MockMvc mockMvc;
-
-	@Autowired
-	protected ObjectMapper objectMapper;
-
-	@Autowired
-	protected WebApplicationContext webApplicationContext;
-
-	@BeforeEach
-	void setUp() {
-		this.mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-			.addFilters(new CharacterEncodingFilter("UTF-8", true))  // 필터 추가
-			.alwaysDo(print())
-			.build();
-	}
-}
